src/requirements.txt
python-dotenv
requests
pymysql
pyEventLogger
beautifulsoup4
pytz
arrow
quickchart.io
fastapi
uvicorn
psycopg[binary]
